# api/openapi.yaml
openapi: 3.1.0
info:
  title: BrotherBand API (Cygnus)
  version: "1.0.0"
  description: |
    Backend HTTP contract for the BrotherBand minimalist social network (Cygnus iteration).

    ## Authentication

    Cookie-based. A successful `POST /v1/auth/login` or `POST /v1/auth/register`
    sets `bb_session` (httpOnly JWT, 30-day) and `bb_csrf` (readable, 32-byte random).
    Every state-changing request must echo the `bb_csrf` value back as the
    `X-CSRF-Token` header.

    All timestamps are RFC 3339. All IDs are UUIDv4 strings.

    ## Error model

    Every non-2xx response uses the `Error` schema: a stable `code`, a
    user-safe `message`, the `requestId`, and — on validation failures —
    a structured `details` object. **Switch on `code`, never on `message`.**

    Stable `code` registry:

    | code | HTTP | meaning |
    |---|---|---|
    | `user.username_taken` | 409 | username already registered |
    | `user.invalid_credentials` | 401 | login failed (cause never disclosed) |
    | `user.password_too_weak` | 422 | password policy not met |
    | `user.invalid_username` / `_birthdate` / `_secret` / `_status` / `_favorites` | 422 | field-level validation failure |
    | `user.not_found` | 404 | user does not exist |
    | `brotherband.self_request` | 422 | cannot brotherband yourself |
    | `brotherband.already_brothers` | 409 | relationship already exists |
    | `brotherband.request_exists` | 409 | a pending request already exists |
...

    ## Observability

...
servers:
...
paths:
...
  /v1/auth/register:
    post:
      tags: [auth]
      operationId: registerUser
      summary: Register a new user account and start a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered; session and CSRF cookies are set
          headers:
            Set-Cookie:
              description: Sets bb_session (httpOnly) and bb_csrf (readable)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /v1/auth/login:
    post:
      tags: [auth]
      operationId: loginUser
      summary: Authenticate with username + password and start a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated; session and CSRF cookies are set
          headers:
            Set-Cookie:
              description: Sets bb_session (httpOnly) and bb_csrf (readable)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/brothers:
    get:
      tags: [brotherband]
      operationId: listBrothers
      summary: List all confirmed brothers of the authenticated user
      responses:
        "200":
          description: List of brothers
          content:
            application/json:
              schema:
                type: object
                required: [brothers]
                properties:
                  brothers:
                    type: array
                    items:
                      $ref: "#/components/schemas/BrotherSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/brothers/{brotherId}:
    get:
      tags: [brotherband]
      operationId: getBrotherProfile
      summary: Get the public profile of a confirmed brother
      parameters:
        - $ref: "#/components/parameters/BrotherId"
      responses:
        "200":
          description: Brother profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrotherProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [brotherband]
      operationId: cutBrotherband
      summary: Cut the brotherband with another user (mutual unfriend)
      parameters:
        - $ref: "#/components/parameters/BrotherId"
      responses:
        "204":
          description: Brotherband cut
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/conversations:
    get:
      tags: [messages]
      operationId: listConversations
      summary: List the authenticated user's conversations (one per brother)
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: object
                required: [conversations]
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: "#/components/schemas/ConversationSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  parameters:
    BrotherId:
      in: path
      name: brotherId
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: bb_session
    CsrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
...
    Error:
      type: object
      required: [code, message]
      description: |
        The uniform error envelope returned by every non-2xx response.

        - `code` is a stable, machine-readable identifier — switch on this,
          never on `message`. The full registry is documented in the API
          description.
        - `message` is a short human-readable summary, safe to surface to
          end users. It never contains stack traces or internal detail.
        - `requestId` echoes the `X-Request-ID` response header. Always
          include it in bug reports — it correlates the client failure with
          the exact server-side structured log line.
        - `details` is present on validation failures. For a single bad
          field it is `{ "field": "...", "reason": "..." }`; for a batch it
          is `{ "fields": [ { "field": "...", "reason": "..." }, ... ] }`.
      properties:
        code:
          type: string
          description: Stable machine-readable error code (see registry in the API description).
          examples: [user.username_taken, message.invalid_body, unauthenticated, internal_error]
        message:
          type: string
          description: Short, user-safe human-readable summary.
        requestId:
          type: string
          description: Correlation id; also returned in the X-Request-ID header.
        details:
          type: object
          description: Structured, error-specific payload. Present on 422 validation failures.
          properties:
            field:
              type: string
              description: The offending field (single-field validation failure).
            reason:
              type: string
              description: Why the field was rejected.
            fields:
              type: array
              description: Multiple field failures, when more than one was detected.
              items:
                type: object
                required: [field, reason]
                properties:
                  field:
                    type: string
                  reason:
                    type: string
          additionalProperties: true
    RegisterRequest:
      type: object
      required: [username, password, birthdate, secret, status, favorites]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: "^[a-zA-Z0-9_-]+$"
        password:
          type: string
          minLength: 8
          maxLength: 128
        birthdate:
          type: string
          format: date
        secret:
          type: string
          description: Personal secret revealed to a user only when they accept this user's brotherband request
          minLength: 1
          maxLength: 280
        status:
          type: string
          minLength: 1
          maxLength: 280
        favorites:
          type: array
          minItems: 5
          maxItems: 5
          items:
            type: string
            minLength: 1
            maxLength: 80
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
...
    UserProfile:
      type: object
      required: [id, username, status, favorites, registeredAt]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        status:
          type: string
        favorites:
          type: array
          items:
            type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
        registeredAt:
          type: string
          format: date-time
    BrotherSummary:
      type: object
      required: [id, username, status]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        status:
          type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
        becameBrothersAt:
          type: string
          format: date-time
    BrotherProfile:
      allOf:
        - $ref: "#/components/schemas/BrotherSummary"
        - type: object
          required: [favorites]
          properties:
            favorites:
              type: array
              items:
                type: string
...
    ConversationSummary:
      type: object
      required: [conversationId, brother, lastMessageAt]
      properties:
        conversationId:
          type: string
          format: uuid
        brother:
          $ref: "#/components/schemas/BrotherSummary"
        lastMessageAt:
          type: string
          format: date-time
          nullable: true
        lastMessagePreview:
          type: string
          nullable: true
        unreadCount:
          type: integer
...
  headers:
    XRequestID:
      description: |
        Per-request correlation id. Generated by the server if the client
        did not supply one on the request. Echoed on every response (success
        and error) and mirrored in the `requestId` field of error bodies.
      schema:
        type: string
      required: true
  responses:
    Unauthorized:
      description: Missing or invalid session
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but not allowed (e.g. not a participant)
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource does not exist
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource state conflicts with the request
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnprocessableEntity:
      description: |
        Validation failed. The `details` object carries the offending
        `field` and `reason` (or a `fields` array for batch failures).
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
security:
  - SessionCookie: []